Added dynamic posting feature for reviews and ratings which fixes #10
Both the review and ratings can be submitted together with one button. The review and ratings are updated to the same new database.
