+++ screens/ItemScreen.js
@@ -1,10 +1,10 @@
 import React, { useEffect, useState } from "react";
 import { Text, View, FlatList, StyleSheet, Image, TouchableOpacity, Alert } from "react-native";
-import ListReview from "../components/ListReview";
 import { supabase } from "../lib/supabase";
 
 //Functionality to fetch image from the database
 //{item_id: route.params.item_id} in review section. Be able to pass the data to the ReviewScreen
+//Unable to add image inside
 
 export const ItemScreen = ( {navigation, route} ) => {
   const [review, setReview] = useState("");
@@ -23,32 +23,37 @@ export const ItemScreen = ( {navigation, route} ) => {
     // console.log(all_reviews);
     // console.log(error);
   };
-
   useEffect(() => {
     LoadReview();
   });
+
   return (
     <View>
       {/* Post a review button */}
       <Text style={styles.headertext}> {route.params.item_name} </Text>
-      <TouchableOpacity style={styles.button} onPress={() => navigation.navigate("Review")}>
+      <TouchableOpacity style={styles.button} onPress={() => navigation.navigate("Review", {item_id: route.params.item_id})}>
         <Text style={styles.buttontext}> Post a review! </Text>
       </TouchableOpacity>
 
-      {/* Image of food item */}
+      {/* 
       <View style={styles.image}>
         <Image
           source={require("../assets/img_test.jpg")}
           style={{ width: 200, height: 200 }}
         />
       </View>
+      Image of food item. Unable to implement this at the moment*/}
 
-      {/* From ListReview component */}
+      {/* Header for reviews section */}
+      <Text style={styles.subheadertext}>Reviews:</Text>
+      
+      {/* Listing of reviews and ratings */}
       <FlatList
+        contentContainerStyle={{ paddingBottom: 200 }}
         keyExtractor={(item) => item.review_id}
         data={review}
         renderItem={({ item }) => (
-            <Text style={styles.item}> 
+            <Text style={styles.reviewtext}> 
               Review: {item.review} {'\n'} 
               Taste: {item.taste} {'\n'}
               Value for money: {item.money} 
@@ -66,21 +71,7 @@ const styles = StyleSheet.create({
   headertext: {
     color: "#A7BC5B",
     fontSize: 28,
-  },
-  innertext: {
-    fontSize: 16,
-    paddingHorizontal: 20,
-    paddingVertical: 30,
-  },
-  image: {
-    alignSelf: "center",
-  },
-  item: {
-    alignItems: "center",
-    backgroundColor: "#8DA242",
-    padding: 20,
-    marginVertical: 10,
-    marginHorizontal: 10,
+    fontWeight: 'bold'
   },
   button: {
     alignItems: "center",
@@ -90,6 +81,23 @@ const styles = StyleSheet.create({
     marginHorizontal: 10,
   },
   buttontext: {
-    fontSize: 20
+    fontSize: 20,
+    fontWeight: 'bold'
+  },
+  subheadertext: {
+    fontSize: 25,
+    paddingHorizontal: 10,
+    fontWeight: 'bold',
+    color: "#A7BC5B"
+  },
+  image: {
+    alignSelf: "center",
+  },
+  reviewtext: {
+    alignItems: "center",
+    backgroundColor: "#8DA242",
+    padding: 15,
+    marginVertical: 10,
+    marginHorizontal: 10,
   }
 });
